Move ql.ts to /common
