+++ extensions/ql-vscode/src/common/ql.ts
