+++ extensions/ql-vscode/src/databases/qlpack.ts
@@ -3,7 +3,7 @@ import { glob } from "glob";
 import { basename } from "path";
 import { load } from "js-yaml";
 import { readFile } from "fs-extra";
-import { getQlPackPath } from "../pure/ql";
+import { getQlPackPath } from "../common/ql";
 import { CodeQLCliServer, QlpacksInfo } from "../codeql-cli/cli";
 import { extLogger } from "../common/logging/vscode";
 import { getOnDiskWorkspaceFolders } from "../common/vscode/workspace-folders";
