+++ extensions/ql-vscode/src/language-support/contextual/query-resolver.ts
@@ -26,7 +26,7 @@ import { CancellationToken } from "vscode";
 import { ProgressCallback } from "../../common/vscode/progress";
 import { CoreCompletedQuery, QueryRunner } from "../../query-server";
 import { redactableError } from "../../common/errors";
-import { QLPACK_FILENAMES } from "../../pure/ql";
+import { QLPACK_FILENAMES } from "../../common/ql";
 import { telemetryListener } from "../../common/vscode/telemetry";
 
 export async function qlpackOfDatabase(
