+++ extensions/ql-vscode/src/variant-analysis/run-remote-query.ts
@@ -33,7 +33,7 @@ import {
   FALLBACK_QLPACK_FILENAME,
   QLPACK_FILENAMES,
   QLPACK_LOCK_FILENAMES,
-} from "../pure/ql";
+} from "../common/ql";
 import { QueryLanguage } from "../common/query-language";
 import { tryGetQueryMetadata } from "../codeql-cli/query-metadata";
 import { askForLanguage, findLanguage } from "../codeql-cli/query-language";
