+++ extensions/ql-vscode/test/unit-tests/pure/ql.test.ts
@@ -2,7 +2,7 @@ import { join } from "path";
 import { dirSync } from "tmp-promise";
 import { DirResult } from "tmp";
 import { writeFile } from "fs-extra";
-import { getQlPackPath } from "../../../src/pure/ql";
+import { getQlPackPath } from "../../../src/common/ql";
 
 describe("getQlPackPath", () => {
   let tmpDir: DirResult;
